requirements.txt
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0 
pyperclip>=1.9.0